Añadiendo registro de importaciones en el linker. Obteniendo strings de las importaciones y añadiendolos a la metaseccion de strings. Añadiendo campos en el header del formato VELB para añadir dos nuevas futuras tablas proximamente

+++ include/emmit/struct_context.h
@@ -380,7 +380,7 @@ namespace Assembly::Bytecode {
          * "kernel32.dll:GetTickCount". el valor es el indice de la entrada
          * dentro de la tabla "import_table".
          */
-        std::unordered_map<std::string, uint32_t> import_lookup;
+        std::unordered_map<std::string, uint64_t> import_lookup;
 
         /**
          * espacio de direcciones key(nombre): valor(espacio)
